+++ include/LLGL/Backend/CommandBuffer.Blitting.inl
@@ -11,7 +11,7 @@ virtual void UpdateBuffer(
     LLGL::Buffer&                   dstBuffer,
     std::uint64_t                   dstOffset,
     const void*                     data,
-    std::uint16_t                   dataSize
+    std::uint64_t                   dataSize
 ) override final;
 
 virtual void CopyBuffer(
