+++ sources/Renderer/Metal/Command/MTDirectCommandBuffer.mm
@@ -118,7 +118,7 @@ void MTDirectCommandBuffer::UpdateBuffer(
     Buffer&         dstBuffer,
     std::uint64_t   dstOffset,
     const void*     data,
-    std::uint16_t   dataSize)
+    std::uint64_t   dataSize)
 {
     auto& dstBufferMT = LLGL_CAST(MTBuffer&, dstBuffer);
